refactor: terms
